move docgen to internal
