+++ internal/docgen/parser.go
