+++ internal/docgen/process.go
