+++ internal/docgen/reader.go
