+++ internal/docgen/serializer.go
